Fix broken table in user_guide
The first row in the table in section 3 (Terminology) has an unordered
list (<ul>) with list items (<li>) in it. Although closing <li>
elements is optional in HTML, and most browsers will render list
items properly even without closing them, Jekyll seems to be stricter
and breaks the table the list was placed in, making sections 3 and
onwards unreadable.

This patch closes the <li> tags with corresponding </li> tags, making
the user_guide page readable again.

Signed-off-by: Allon Mureinik <amureini@redhat.com>